# swagger.yaml
openapi: 3.0.3
info:
  title: Girls Got Game API
  description: Basketball training application API for mobile and web clients
  version: 2.0.0
  contact:
    name: Girls Got Game API Support
    email: support@girlsgotgame.app
servers:
  - url: http://localhost:3001/api
    description: Development server
  - url: https://api.girlsgotgame.app/api
    description: Production server
security:
  - cookieAuth: []
paths:
...
  /games:
    get:
      tags:
        - Games
      summary: Get all games
      description: Retrieve all games (public endpoint)
      responses:
        '200':
          description: Games retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Game'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - Games
      summary: Create game
      description: Create a new game (admin only)
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGame'
      responses:
        '201':
          description: Game created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Game'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /games/my-games:
    get:
      tags:
        - Games
      summary: Get user's games
      description: Get games where the authenticated user or their children participated
      security:
        - cookieAuth: []
      responses:
        '200':
          description: User games retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Game'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /games/{gameId}:
    get:
      tags:
        - Games
      summary: Get game by ID
      description: Retrieve a specific game by ID
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Game retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameWithDetails'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    patch:
      tags:
        - Games
      summary: Update game
      description: Update an existing game (admin only)
      security:
        - cookieAuth: []
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateGame'
      responses:
        '200':
          description: Game updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Game'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags:
        - Games
      summary: Delete game
      description: Delete a game (admin only)
      security:
        - cookieAuth: []
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Game deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /admin/users:
    get:
      tags:
        - Admin
      summary: Get all users
      description: Retrieve all users (admin only)
      security:
        - cookieAuth: []
      parameters:
        - name: search
          in: query
          schema:
            type: string
          description: Search users by email or name
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Users retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AdminUser'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /admin/admins:
    get:
      tags:
        - Admin
      summary: Get all admin users
      description: Retrieve all users with admin permissions
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Admin users retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AdminUser'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /admin/grant-admin:
    post:
      tags:
        - Admin
      summary: Grant admin permissions
      description: Grant admin permissions to a user
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: string
                  format: uuid
                email:
                  type: string
                  format: email
              oneOf:
                - required: [userId]
                - required: [email]
      responses:
        '200':
          description: Admin permissions granted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  user:
                    $ref: '#/components/schemas/AdminUser'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /admin/revoke-admin:
    post:
      tags:
        - Admin
      summary: Revoke admin permissions
      description: Revoke admin permissions from a user
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: string
                  format: uuid
              required:
                - userId
      responses:
        '200':
          description: Admin permissions revoked successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  user:
                    $ref: '#/components/schemas/AdminUser'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /admin/check-admin:
    get:
      tags:
        - Admin
      summary: Check admin status
      description: Check if current user has admin permissions
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Admin status retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  isAdmin:
                    type: boolean
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # Media Endpoints
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
  schemas:
...
    AdminUser:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        name:
          type: string
          nullable: true
        role:
          type: string
          enum: [player, parent]
        isAdmin:
          type: boolean
        isVerified:
          type: boolean
        totalPoints:
          type: integer
        createdAt:
          type: string
          format: date-time
...
    Game:
      type: object
      properties:
        id:
          type: string
          format: uuid
        teamName:
          type: string
        isHome:
          type: boolean
        opponentTeam:
          type: string
        gameDate:
          type: string
          format: date-time
        homeScore:
          type: integer
          nullable: true
        awayScore:
          type: integer
          nullable: true
        notes:
          type: string
          nullable: true
        status:
          type: string
          enum: [scheduled, in_progress, completed, cancelled]
        isSharedToFeed:
          type: boolean
          default: false
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    GameWithDetails:
      allOf:
        - $ref: '#/components/schemas/Game'
        - type: object
          properties:
            players:
              type: array
              items:
                $ref: '#/components/schemas/GamePlayer'
            activities:
              type: array
              items:
                $ref: '#/components/schemas/GameActivity'
    GamePlayer:
      type: object
      properties:
        id:
          type: string
          format: uuid
        gameId:
          type: string
          format: uuid
        userId:
          type: string
          nullable: true
        manualPlayerId:
          type: string
          format: uuid
          nullable: true
        playerName:
          type: string
        jerseyNumber:
          type: integer
          nullable: true
    GameActivity:
      type: object
      properties:
        id:
          type: string
          format: uuid
        gameId:
          type: string
          format: uuid
        activityType:
          type: string
        description:
          type: string
        metadata:
          type: object
          nullable: true
        createdAt:
          type: string
          format: date-time
    CreateGame:
      type: object
      required:
        - teamName
        - isHome
        - opponentTeam
        - gameDate
      properties:
        teamName:
          type: string
        isHome:
          type: boolean
        opponentTeam:
          type: string
        gameDate:
          type: string
          format: date-time
        notes:
          type: string
    UpdateGame:
      type: object
      properties:
        teamName:
          type: string
        isHome:
          type: boolean
        opponentTeam:
          type: string
        gameDate:
          type: string
          format: date-time
        homeScore:
          type: integer
        awayScore:
          type: integer
        notes:
          type: string
        status:
          type: string
          enum: [scheduled, in_progress, completed, cancelled]
        isSharedToFeed:
          type: boolean
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Invalid data provided
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Authentication required
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Insufficient permissions
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Resource not found
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Internal server error
...